image-deblurring/requirements.txt
numpy
pillow
torchvision
onnx
onnxruntime    
streamlit